website/requirements.txt
# Requirements file
# Libraries worden automatisch bij deployment geinstalleerd
# De opgeven versie is optioneel
# Mocht een requirement een foutmelding geven, probeer dan het versienummer weg te laten
# Dit kan het gevolg zijn van een andere Python versie
Flask==1.0.3
Jinja2>=2.10.1
Werkzeug==0.15.4
mysql-connector==2.2.9
biopython==1.73